# Conflicts:
#	.gitignore
#	application/api/controllers/AutocompleteController.php
#	application/api/controllers/CollectionsController.php
#	application/api/controllers/ConceptController.php
#	application/api/controllers/FindConceptsController.php
#	application/api/controllers/InstitutionsController.php
#	application/api/controllers/RelationController.php
#	application/api/views/scripts/concept/get.phtml
#	application/api/views/scripts/index/index.phtml
#	application/editor/controllers/ConceptController.php
#	application/editor/forms/Concept/FormToConcept.php
#	application/oai-pmh/controllers/IndexController.php
#	library/OpenSKOS/Db/Table/Row/Tenant.php
#	library/OpenSKOS/Db/Table/Row/User.php
#	library/OpenSkos2/Api/Concept.php
#	library/OpenSkos2/Api/Response/Detail/JsonResponse.php
#	library/OpenSkos2/Api/Response/Detail/JsonpResponse.php
#	library/OpenSkos2/Api/Response/ResultSet/JsonResponse.php
#	library/OpenSkos2/Api/Response/ResultSet/JsonpResponse.php
#	library/OpenSkos2/Api/Response/ResultSet/RdfResponse.php
#	library/OpenSkos2/Api/Response/ResultSetResponse.php
#	library/OpenSkos2/Api/Transform/DataArray.php
#	library/OpenSkos2/Api/Transform/DataRdf.php
#	library/OpenSkos2/Bridge/EasyRdf.php
#	library/OpenSkos2/Concept.php
#	library/OpenSkos2/ConceptManager.php
#	library/OpenSkos2/Converter/Text.php
#	library/OpenSkos2/EasyRdf/Serialiser/RdfXml/OpenSkos.php
#	library/OpenSkos2/Export/Command.php
#	library/OpenSkos2/Export/Message.php
#	library/OpenSkos2/Export/Serialiser.php
#	library/OpenSkos2/Export/Serialiser/FormatAbstract.php
#	library/OpenSkos2/FieldsMaps.php
#	library/OpenSkos2/Import/Command.php
#	library/OpenSkos2/Rdf/Object.php
#	library/OpenSkos2/Rdf/Resource.php
#	library/OpenSkos2/Rdf/ResourceManager.php
#	library/OpenSkos2/Rdf/Uri.php
#	library/OpenSkos2/SkosXl/LabelManager.php
#	library/OpenSkos2/Solr/Document.php
#	library/OpenSkos2/Solr/ResourceManager.php
#	library/OpenSkos2/Tenant.php
#	library/OpenSkos2/Validator/Collection.php
#	library/OpenSkos2/Validator/Resource.php
#	package.json
#	tools/migrate.php

+++ library/OpenSkos2/Rdf/ResourceManagerWithSearch.php
@@ -29,6 +29,7 @@ use OpenSkos2\Rdf\ResourceCollection;
 
 class ResourceManagerWithSearch extends ResourceManager
 {
+
     /**
      * @var \OpenSkos2\Solr\ResourceManager
      */
@@ -71,9 +72,20 @@ class ResourceManagerWithSearch extends ResourceManager
     public function insert(Resource $resource)
     {
         parent::insert($resource);
-        $this->solrResourceManager->insert($resource);
+
+        if ($resource->getType()->getUri() === Concept::TYPE) {
+            $specs = $this->fetchConceptSpec($resource);
+            if (count($specs) < 1) {
+                throw new \Exception("Cannot fetch spec for the concept {$resource->getUri()}");
+            }
+            foreach ($specs as $spec) {
+                $resource->setProperty(OpenSkosNamespace::TENANT, new Uri($spec['tenanturi']));
+                $resource->setProperty(OpenSkosNameSpace::SET, new Uri($spec['seturi']));
+            }
+            $this->solrResourceManager->insert($resource);
+        }
     }
-    
+
     /**
      * @param \OpenSkos2\Rdf\ResourceCollection $resourceCollection
      * @throws ResourceAlreadyExistsException
@@ -83,7 +95,7 @@ class ResourceManagerWithSearch extends ResourceManager
         parent::insertCollection($resourceCollection);
         $this->solrResourceManager->insertCollection($resourceCollection);
     }
-    
+
     /**
      * Deletes and then inserts the resourse.
      * @param \OpenSkos2\Rdf\Resource $resource
@@ -91,9 +103,10 @@ class ResourceManagerWithSearch extends ResourceManager
     public function replace(Resource $resource)
     {
         parent::replace($resource);
-        
-        $this->solrResourceManager->delete($resource);
-        $this->solrResourceManager->insert($resource);
+        if ($resource->getType()->getUri() === Concept::TYPE) {
+            $this->solrResourceManager->delete($resource);
+            $this->solrResourceManager->insert($resource);
+        }
     }
 
     /**
@@ -120,15 +133,15 @@ class ResourceManagerWithSearch extends ResourceManager
     public function search($query, $rows = 20, $start = 0, &$numFound = 0, $sorts = null)
     {
         $filterQueries = null;
-        
+
         if (!empty($this->resourceType)) {
             $filterQueries = [
                 'rdfTypeFilter' => 's_rdfType:"' . $this->resourceType . '"'
             ];
         }
-        
+
         return $this->fetchByUris(
-            $this->solrResourceManager->search($query, $rows, $start, $numFound, $sorts, $filterQueries)
+                $this->solrResourceManager->search($query, $rows, $start, $numFound, $sorts, $filterQueries)
         );
     }
 
@@ -139,4 +152,5 @@ class ResourceManagerWithSearch extends ResourceManager
     {
         $this->solrResourceManager->commit();
     }
+
 }
